Real Time Lane Detection for Autonomous Vehicals
In this project, I would like to reproduce and modify algorithms that are available in the research papers - "Lane Detection for Autonomous Vehicles using OpenCV Library" and "Real-time Lane detection and Motion Planning in Raspberry Pi and Arduino for an Autonomous Vehicle Prototype". I will be writing an algorithm that can detect lanes from a captured image, video, and real-time. I will be using various methodologies like the Canny Edge Detection Technique, Gaussian blur, Region of Interest, and Hough Transform.